python/subscribe/requirements.txt
remotivelabs-broker == 0.1.21
thrift
apache-iotdb